final work

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,15 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+
+    const authenticatedUserId = req.session.userId;
+
     try {
-        const notes = await NoteModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+        const notes = await NoteModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(notes);
     } catch (error) {
         next(error);
@@ -13,14 +18,19 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 }
 
 export const getNote: RequestHandler =async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
     const noteId = req.params.noteId;
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id.");
         }
         const note = await NoteModel.findById(noteId).exec();
         if(!note) {
             throw createHttpError(404, "Note not found.")
+        }
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note.");
         }
         res.status(200).json(note);
     } catch (error) {
@@ -36,12 +46,15 @@ interface CreateNoteBody {
 export const createNote: RequestHandler<unknown, unknown, CreateNoteBody, unknown> = async (req, res, next) => {
     const title = req.body.title;
     const text = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if (!title) {
             throw createHttpError(400, "Note must have a title.");
         }
         const newNote = await NoteModel.create({
+            userId: authenticatedUserId,
             title: title,
             text: text,
         });
@@ -64,8 +77,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
     const noteId = req.params.noteId;
     const newTitle = req.body.title;
     const newText = req.body.text;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id.");
         }
@@ -80,6 +95,10 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
             throw createHttpError(404, "Note not found.");
         }
 
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note.");
+        }
+
         note.title = newTitle;
         note.text = newText;
 
@@ -92,9 +111,11 @@ export const updateNote: RequestHandler<UpdateNoteParams, unknown, UpdateNoteBod
 }
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
-    const noteId = req.params.noteId
+    const noteId = req.params.noteId;
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         if(!mongoose.isValidObjectId(noteId)) {
             throw createHttpError(400, "Invalid note id.");
         }
@@ -103,6 +124,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
         if(!note) {
             throw createHttpError(404, "Note not found.");
+        }
+
+        if(!note.userId.equals(authenticatedUserId)) {
+            throw createHttpError(401, "You cannot access this note.");
         }
 
         await note.deleteOne();
